requirements.txt
nba_api
pandas
flask
line-bot-sdk
python-dotenv
